GPT-236 Check for SLD tag and apply filter params to SLD request
Some ArcGIS services do not like the XML declaration in the SLD_BODY
parameter, so we must leave it off. Therefore the portal needs to
check whether there is either a <?xml or <StyledLayerDescriptor partial
tag. We also apply the filterer parameters here to the SLD proxy request,
following what is done in FeatureWithMapRenderer.js