AUS-2664 FilterForm constructor called too many times when "Add Layer to Map" pressed
Every time the "loading" attribute in the 'KnownLayer' model is
changed, it triggers the constructor for the FilterForm via
'restoreRowContainer()' in "RowExpanderContainer.js" etc. so I removed it,
as it does not seem to serve any purpose, other than just to say "I'm
loading".
The "active" attribute is already being set/unset and this is enough to
trigger the constructor. Setting or unsetting the "loading" attribute is
redundant.