Add visibile to Layer model
Added the visible flag to the model for Layer, such that a store update event would change the icon in the ActiveLayerPanel. This means there is a duplication of visible flags in the Layer class, but I left the original there in case othe rportals depend on that flag.

The visible flag outside the model may be a candidate for deprecation, as noted in comments.